Save memory by storing colors as uint8 instead of float32 (#4210)
Memory savings per vertex:

```
  ┌───────────────────────┬──────────┬──────────┬───────────┐
  │        Struct         │  Before  │  After   │ Reduction │
  ├───────────────────────┼──────────┼──────────┼───────────┤
  │ PointPainter::Data    │ 28 bytes │ 16 bytes │ 43%       │
  ├───────────────────────┼──────────┼──────────┼───────────┤
  │ MeshPainter::Data     │ 36 bytes │ 28 bytes │ 22%       │
  ├───────────────────────┼──────────┼──────────┼───────────┤
  │ LinePainter::Data     │ 56 bytes │ 32 bytes │ 43%       │
  ├───────────────────────┼──────────┼──────────┼───────────┤
  │ TrianglePainter::Data │ 84 bytes │ 48 bytes │ 43%       │
  └───────────────────────┴──────────┴──────────┴───────────┘
  ```

+++ src/colmap/ui/mesh_painter.h
@@ -31,6 +31,7 @@
 
 #include <QtCore>
 #include <QtOpenGL>
+#include <cstdint>
 
 namespace colmap {
 
@@ -47,14 +48,14 @@ class MeshPainter {
          float nx,
          float ny,
          float nz,
-         float r,
-         float g,
-         float b)
+         uint8_t r,
+         uint8_t g,
+         uint8_t b)
         : px(px), py(py), pz(pz), nx(nx), ny(ny), nz(nz), r(r), g(g), b(b) {}
 
     float px, py, pz;
     float nx, ny, nz;
-    float r, g, b;
+    uint8_t r, g, b;
   };
 
   void Setup();
